src/backend/requirements.txt
fpdf2
psycopg2-binary
boto3